Cleaned up some hanging code
