+++ src/components/Dashboard.js
@@ -3,11 +3,7 @@ import React from 'react';
 import DashboardChart from './DashboardChart';
 import NavMap from './NavMap';
 
-import { useIncidents } from '../state/query_hooks/useIncidents';
-
 const Dashboard = () => {
-  const incidentsQuery = useIncidents();
-
   return (
     <div
       className="dashboard-container"
