+++ src/components/DashboardChart.js
@@ -4,7 +4,7 @@ import { PieData } from './PieData';
 import { useIncidents } from '../state/query_hooks/useIncidents';
 
 const DashboardChart = () => {
-  const [limit, setLimit] = useState(100);
+  const [limit] = useState(100);
 
   let pie_data = PieData(limit);
   const incidentsQuery = useIncidents();
@@ -12,8 +12,6 @@ const DashboardChart = () => {
   if (incidentsQuery?.data?.length > 0) {
     return (
       <div className="dash-chart-container">
-        {/* ⬇️ pushes legend below footer -- needs a solution [@ nic] */}
-        {/* <h1 style={{ textAlign: 'center' }}> Most Common Incident Types</h1> */}
         <PieChart data={pie_data} />
       </div>
     );
