+++ src/components/NavDashboard.js
@@ -4,7 +4,6 @@ import { Link } from 'react-router-dom';
 const NavDashboard = () => {
   return (
     <div className="dashboard-nav">
-      {/* <Link to="/dashboard">Dashboard</Link> */}
       <Link to="/dashboard-chart">Explore</Link>
     </div>
   );
