--- src/database/data1.json
@@ -1,14 +1,0 @@
-{
-  "data": [
-    {
-      "id": 0,
-      "date_text": "June 3rd",
-      "text": "Police throw tear-gas at protesters on a bridge.",
-      "tags_str": "Projectiles",
-      "LATITUDE": 29.963071000000006,
-      "LONGITUDE": -90.160953,
-      "Link1": "https://twitter.com/misaacstein/status/1268381797081022464",
-      "Link2": "https://twitter.com/ckm_news/status/1268382403367763970"
-    }
-  ]
-}
