--- src/utils copy/axiosWithAuth.js
@@ -1,13 +1,0 @@
-import axios from 'axios';
-
-export const axiosWithAuth = () => {
-  const token = localStorage.getItem('token');
-  return axios.create({
-    headers: {
-      Authorization: token,
-    },
-    //        baseURL: 'https://expat-backend.herokuapp.com'
-  });
-};
-
-export default axiosWithAuth;
